Add network_tester application as a new role
Move the previous standalone kritis3m_network_tester into this
application as a new role. This was necessary to support Zephyr in the
future. This also simplifies code management, as there is no code
duplication anymore between both repositories.

To incorporate the new application, the general executable has been
renamed to `kritis3m_tls`. This also better matches the repo name and
the other existing roles.